+++ Sources/VCore/Presentation/SwiftUI/Progress View (With Parameters)/View+ProgressView.swift
@@ -21,15 +21,85 @@ extension View {
         parameters: ProgressViewParameters?
     ) -> some View {
         self
+            .modifier(
+                ProgressViewModifier(
+                    parameters: parameters
+                )
+            )
+    }
+}
+
+private struct ProgressViewModifier: ViewModifier {
+    // MARK: Properties - Parameters
+    private let parameters: ProgressViewParameters?
+    
+    @State private var isVisible: Bool = false
+    
+    // MARK: Subscriptions
+    @State private var visibilityTask: Task<Void, Never>?
+    
+    // MARK: Initializers
+    init(
+        parameters: ProgressViewParameters?
+    ) {
+        self.parameters = parameters
+    }
+    
+    // MARK: Body
+    func body(content: Content) -> some View {
+        content
             .blocksHitTesting(parameters?.isInteractionEnabled == false)
             .overlay {
-                if let parameters {
+                if
+                    let parameters,
+                    isVisible
+                {
                     ProgressView()
                         .progressViewStyle(.circular)
                         .scaleEffect(parameters.scalingFactor ?? 1)
                         .tint(parameters.color)
                 }
             }
+        
+            .onChange(of: parameters.visibilityData, initial: true) { (_, newValue) in
+                visibilityTask?.cancel()
+                visibilityTask = nil
+                
+                if newValue.isVisible {
+                    if let delay: Duration = newValue.delay {
+                        visibilityTask = Task {
+                            do {
+                                try await Task.sleep(for: delay)
+                            } catch {
+                                return
+                            }
+                            
+                            isVisible = true
+                        }
+                        
+                    } else {
+                        isVisible = true
+                    }
+                    
+                } else {
+                    isVisible = false
+                }
+            }
+    }
+}
+
+// Needed, as `ProgressViewParameters` cannot be equatable
+nonisolated private struct VisibilityData: Equatable {
+    let isVisible: Bool
+    let delay: Duration?
+}
+
+nonisolated extension Optional where Wrapped == ProgressViewParameters {
+    fileprivate var visibilityData: VisibilityData {
+        .init(
+            isVisible: self != nil,
+            delay: self?.appearanceDelay
+        )
     }
 }