+++ CHANGELOG.md
@@ -6,6 +6,10 @@ Core - Utilities
 
 - `SendableTask` is added
 
+Presentation
+
+- `appearanceDelay` is added to `ProgressViewParameters`
+
 ### [8.1.0(114)](https://github.com/VakhoKontridze/VCore/releases/tag/8.1.0) — 2026 05 04
 
 Core - Macros